decimal datatype to int for price

+++ backend/db/migrations/20220404173635-create-spot.js
@@ -35,7 +35,7 @@ module.exports = {
       },
       price: {
         allowNull: false,
-        type: Sequelize.DECIMAL(3,2)
+        type: Sequelize.INTEGER
       },
 
       createdAt: {
